Clean up the utilities and function methodolgy. - Still tons of documentation to do - Move the CppTokenizer to the utils - Fix for using namespace ...; lookup - Fix for nested scope bits
