+++ CppToolkit.sublime-settings
@@ -4,4 +4,5 @@
     // ----------------- [WARNING] -----------------
     // Always make changes in the user settings
     // ----------------- [WARNING] -----------------
+    "foo" : "bar"
 }
