+++ cpp_refactor_commands.py
@@ -9,304 +9,61 @@ import sublime
 import sublime_plugin
 
 from copy import deepcopy
+from .lib.utils import CppTokenizer
 
-class CppTokenizer(object):
+class CppRefactorDetails(object):
     """
-    Utility for building tokens of C++ files. This is by no means complete
-    but forgoes a lot of the nitty gritty to be lean and fast 
+    Class handling the large amount of information surrouding our context aware
+    commands. This is passed to the get_commands(...) function of the concrete
+    _BaseCppCommand(s)
     """
-    DELIMITS = ( '*', '=', '{', '}', '\'', '\"', '(', ')', ';', ':', ' ', '\n', '\t' )
+    def __init__(self, **kwargs):
+        self._view = kwargs.get('view')
+        self._command = kwargs.get('command')
+        self._args = kwargs.get('args')
+        self._pos = kwargs.get('pos')
+        self._current_word = kwargs.get('current_word')
+        self._current_line = kwargs.get('current_line')
+        self._header = kwargs.get('header')
+        self._source = kwargs.get('source')
+        self._current_file_type = kwargs.get('current_file_type')
 
-    def __init__(self, view, start=0, end=None, use_line=None):
-        self._view = view
-        self._current = start
-        if end:
-            self._end = end - self._view.line_height()
-        else:
-            self._end = self._view.layout_extent()[1]
-        self._use_line = use_line
-        self._current_tokens = None
+    @property
+    def view(self):
+        return self._view
 
+    @property
+    def command(self):
+        return self._command
 
-    def _context_line(self, pos: (int, float)):
-        """
-        Get the line of text from our current view based on a position
-        """
-        return self._view.substr(self._view.line(self._view.layout_to_text((0, pos))))        
+    @property
+    def args(self):
+        return self._args
 
+    @property
+    def pos(self):
+        return self._pos
 
-    def _get_tokens(self, line: str) -> list:
-        """
-        Search for additional items to break up our tokens by
-        """
-        tokens = []
-        current = ''
+    @property
+    def current_word(self):
+        return self._current_word
 
-        previous = None
-        line = line.strip()
-        spin = 0
+    @property
+    def current_line(self):
+        return self._current_line
 
-        for i, char in enumerate(line):
-            if spin > 0:
-                spin -= 1
-                continue
+    @property
+    def header(self):
+        return self._header
 
-            if char == '*' and previous == '/':
-                # -- multi-line comment
-                tokens.append('/*')
-                continue
+    @property
+    def source(self):
+        return self._source
 
-            if char == '*' and (i + 1 < len(line)) and line[i+1] == '/':
-                tokens.append('*/')
-                spin = 1
-                continue
-
-            if char not in CppTokenizer.DELIMITS:
-                current += char
-            else:
-                if current:
-                    tokens.append(current)
-                current = ''
-                if char not in (' ', '\n', '\t'):
-                    tokens.append(char)
-            previous = char
-
-        if current:
-            tokens.append(current)
-
-        return tokens
-
-
-    def next(self, **kwargs):
-        """
-        Rather than host the whole buffer in one shot, we just get a
-        line at a time and keep requesting it until we're done
-        """
-
-        # Grab a token list
-        while (self._current_tokens in (None, [])):
-
-            if self._use_line is not None:
-                self._current_tokens = self._get_tokens(self._use_line)
-            else:
-                if self._current > self._end:
-                    # We've made it where we wanted to go
-                    self._current_tokens = None
-                    return None
-
-                toks = self._get_tokens(self._context_line(self._current).strip())
-                self._current += self._view.line_height()
-                if toks:
-                    self._current_tokens = toks
-                    break
-
-        # The active token awaits!
-        current_token = self._current_tokens.pop(0)
-
-        if kwargs.get('in_comment'):
-            return current_token # comments don't validate
-
-        if current_token in ('', ' '):
-            return self.next() # Keep going
-
-        # Basic Validation
-        if current_token.startswith('//'):
-            # Line comment, skip the rest of the line
-            self.skip_line()
-            return self.next()
-
-        if current_token.startswith('/*'):
-            # We have a inner comment, just
-            # spin until we're out of tokens or
-            # we hit the other side of the comment
-            while True:
-                tok = self.next(in_comment=True)
-                if tok is None:
-                    return None
-
-                if tok.endswith('*/'):
-                    # We've hit the end of the comment so whatever comes
-                    # next should be the right bit
-                    return self.next()
-
-        return current_token
-
-    def current_point(self):
-        """
-        :return: sublime point that dictates where in the file we are
-        """
-        # FIXME: Need a better understanding of X column
-        return self._view.layout_to_text((10000, self._current))
-
-    def skip_line(self):
-        """
-        Pass on the rest of our current tokens
-        """
-        self._current_tokens = None
-
-
-    def spin_until(self, char):
-        while True:
-            token = self.next()
-            if token is None:
-                return
-
-            if char in token:
-                return
-
-    @classmethod
-    def ownership_chain(cls, view, at_location):
-        """
-        Build the ownership chain of the currently selected item by
-        identifying the scope we fall into
-        :return: list[list[str(class|struct|namespace), str]]
-        """
-        proc_tokens = ( 'class', 'struct', 'namespace' )
-        izer = cls(view, 0, at_location[1])
-
-        chain = []
-        active_proc = []
-
-        while True:
-            token = izer.next()
-            if token is None:
-                # Nothing left
-                break
-
-            if token in proc_tokens:
-
-                if active_proc:
-                    chain.append(active_proc)
-
-                active_proc = [token, None]
-                while True:
-                    #
-                    # Find the proc name
-                    #
-                    inner_tok = izer.next()
-                    if inner_tok is None or inner_tok.endswith(';'):
-                        #
-                        # We've hit the EOF or a forward declaration,
-                        # let's skip this all together
-                        #
-                        active_proc[1] = None
-                        break;
-
-                    if inner_tok == '{':
-                        #
-                        # We should have found it
-                        #
-                        break
-
-                    if inner_tok == ':':
-                        #
-                        # We have to spin the token until we get to a scope
-                        # opener or a terminator
-                        #
-                        while True:
-                            lower_tok = izer.next()
-                            if lower_tok is None:
-                                active_proc[1] = None
-                                failed = True
-                                break
-
-                            if '{' in lower_tok or ';' in lower_tok:
-                                # We've found the opening to the class
-                                # and we can let the rest of the process
-                                # take place
-                                break
-                        break
-
-                    active_proc = [token, inner_tok]
-
-                if active_proc[1] is None:
-                    # We don't have a class defined
-                    if chain:
-                        active_proc = chain.pop()
-                    else:
-                        active_proc = []
-
-            elif '}' in token:
-                #
-                # We're at the end of a proc scope
-                #
-                if chain:
-                    active_proc = chain.pop()
-                else:
-                    active_proc = []
-
-            elif token == '{':
-                #
-                # The start of a scope that isn't tied to a proc
-                #
-                izer.spin_until('}')
-
-        return (chain + ([active_proc] if active_proc else []))
-
-
-    @classmethod
-    def location_outside(cls, view, root_ownership):
-        """
-        :return: point - location outside of the ending scope of our class, struct,
-        or namespace
-        """
-        izer = cls(view)
-
-        found_proc = False
-        scope_count = 0
-
-        while True:
-            token = izer.next()
-            if token is None:
-                break # Nothing left
-
-            if not found_proc and token == root_ownership[0]:
-                #
-                # We have the right type, now we just need to check if we have the
-                # right name
-                #
-                while True:
-                    inner_tok = izer.next()
-                    if inner_tok is None or inner_tok.endswith(';'):
-                        found_proc = False
-                        break # Not the right one
-
-                    if inner_tok == root_ownership[1]:
-                        #
-                        # We have the proc name, but we still have to make sure
-                        # this isn't a forward declare
-                        #
-                        found_proc = True
-
-                    if inner_tok == '{':
-                        #
-                        # If, by this point, we have found the item, it means
-                        # we're in it's scope, we now just work until we exit
-                        # said scope
-                        #
-                        # However if we haven't found the item, it means we're
-                        # looking at another item
-                        #
-                        break
-
-            if found_proc:
-                #
-                # Now that we know about our type, we need to keep moving until
-                # we find the end of it's scope
-                #
-                if token == '{':
-                    scope_count += 1
-                if token == '}':
-                    if scope_count == 0:
-                        # We've made it!
-                        if root_ownership[0] != 'namespace':
-                            izer.spin_until(';') # Get passed the terminator
-                        return izer.current_point()
-                    else:
-                        scope_count -= 1
-
-        # We couldn't find the end of that scope
-        return None
+    @property
+    def current_file_type(self):
+        return self._current_file_type
+    
 
 
 class _BaseCppRefactorMeta(type):
@@ -363,7 +120,7 @@ class _BaseCppCommand(sublime_plugin.TextCommand, metaclass=_BaseCppRefactorMeta
         return re.sub(regex, _snake, cls.__name__.replace("Command", ''), 0)
 
     @classmethod
-    def get_commands(cls, view, command, args, pos, current_line, header, source):
+    def get_commands(cls, detail):
         pass
 
 
@@ -394,16 +151,17 @@ class CppDeclareInSourceCommand(_BaseCppCommand):
     DECLARE_FORMAT = "{type}{ownership}{method}({source_arguments}){classifiers}"
 
     @classmethod
-    def get_commands(cls, view, command, args, pos, current_line, header, source):
+    def get_commands(cls, detail):
         """
         Check to see if this is a header function of some sort
         TODO: Use the tokenizer to parse the file rather than the regex mess
         """
-        match = cls.HEADER_FUNCTION.match(current_line)
+        view = detail.view
+        match = cls.HEADER_FUNCTION.match(detail.current_line)
         if not match:
             return None
 
-        original_position = pos[:]
+        original_position = detail.pos[:]
         chain = CppTokenizer.ownership_chain(view, original_position)
 
         func_priv = 'default' # public, private, protected, etc (future use)
@@ -423,7 +181,7 @@ class CppDeclareInSourceCommand(_BaseCppCommand):
 
         match_data = match.groupdict()
         match_data.update({
-            "current_line" : current_line,
+            "current_line" : detail.current_line,
             "ownership_chain" : chain,
             "function_priv" : func_priv
         })
@@ -438,9 +196,9 @@ class CppDeclareInSourceCommand(_BaseCppCommand):
         copy_declare.update({ 'in_' : 'clipboard' })
 
         return [
-            ["Declare In {}".format(os.path.basename(source)), 'source_file', source_declare],
+            ["Declare In {}".format(os.path.basename(detail.source)), 'source_file', source_declare],
             # This works in the header_file
-            ["Declare In {}".format(os.path.basename(header)), 'header_file', header_declare],
+            ["Declare In {}".format(os.path.basename(detail.header)), 'header_file', header_declare],
             ["Copy Declaration to Clipboard", 'header_file', copy_declare]
         ]
 
@@ -511,9 +269,11 @@ class CppDeclareInSourceCommand(_BaseCppCommand):
             else:
 
                 # We attempt to declare just outside the highest ownership scope
+                point = None
                 if data['ownership_chain']:
                     point = CppTokenizer.location_outside(self.view, data['ownership_chain'][0])
-                else:
+                
+                if point is None:
                     point = self.view.size()
 
             self.view.insert(edit, point, full_boddy)
